+++ CHANGELOG.md
@@ -40,6 +40,7 @@ All notable changes to this project will be documented here. The format is based
 - Fixed the scene scale factor being ignored when using PCSA-derived muscle sizing (#511)
 - Deleted duplicate `double_pendulum.osim` example file (#512)
 - Added prefixes to various 'subject01' example files (#513)
+- Blacklisted adding `Blankevoort1991Ligament` (it requires a GeometryPath editor to work, #518 #522)
 
 ## [0.3.1] - UNRELEASED